Use boost::filesystem::path::imbue only on Windows
On Linux, the encoding doesn't need to be set to UTF-8, because
that is used already by default. Using the locale broke the minimal
Linux docker container (probably due to some dependency on the
system to find the locale information).

We use the '_WIN32' definition to distinguish between Linux and
Windows (independent of compiler).